requirements.txt
pika~=1.3.2
numpy
requests~=2.32.3
pyyaml
ultralytics
opencv-python-headless==4.9.0.80
redis